Swapping around order

+++ main.py
@@ -22,13 +22,18 @@ async def on_ready():
         i += 1
 
 async def move_message_to_log(message):
-    message_copy = await client.get_channel(config["log_channel_id"]).send(content = message.content, files = [await f.to_file() for f in message.attachments])
+    channel = client.get_channel(config["log_channel_id"])
+    pinned_messages = await channel.pins()
 
-    try:
-        await message_copy.add_reaction(message.reactions[0])
-    except discord.errors.HTTPException:
-        await message_copy.add_reaction("🍑")
-    await message.delete()
+    if message not in pinned_messages:
+        message_copy = await channel.send(content = message.content, files = [await f.to_file() for f in message.attachments])
+
+        try:
+            await message_copy.add_reaction(message.reactions[0])
+        except discord.errors.HTTPException:
+            await message_copy.add_reaction("🍑")
+
+        await message.delete()
 
 @client.event
 async def on_raw_reaction_add(payload):
